// backend/src/app/modules/communication/enums/notification-templates.enum.ts
export enum NotificationTemplate {
  NEW_CHAT_MESSAGE = 'NEW_CHAT_MESSAGE', // to do
  DIRECT_CALL_INVITE = 'DIRECT_CALL_INVITE',
  DIRECT_CALL_MISSED = 'DIRECT_CALL_MISSED',
  ASSESSMENT_INVITATION = 'ASSESSMENT_INVITATION',
  INTERVIEW_SETUP = 'INTERVIEW_SETUP',
  INTERVIEW_FEEDBACK = 'INTERVIEW_FEEDBACK',
  PRESCREEN_INVITATION_QUESTIONNAIRE = 'PRESCREEN_INVITATION_QUESTIONNAIRE',
  INTERVIEW_SCHEDULED_CONFIRMATION = 'INTERVIEW_SCHEDULED_CONFIRMATION',
  CONFIRMATION_INTERVIEW_SCHEDULED = 'CONFIRMATION_INTERVIEW_SCHEDULED' //old
  // OFFER = 'OFFER',
  // HIRED = 'HIRED',
  // REJECTED = 'REJECTED',
}